Add some JLab Farm script skeletons
Some skeleton scripts which can be modified and fleshed out for usage on the JLab iFarm.

One is straightforwardly, a script which created and submits jobs (JLab_Farming.sh). This submits jobs which run the other shell script, JLab_Farming_Job.sh.

Also included is a simple script to initialise the environment within eic-shell for the job.